worktually_v3_api/pyproject.toml
[tool.poetry]
name = "worktually-v3-api"
version = "0.1.0"
description = ""
authors = ["Mustafa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
amqp = "                         5.2.0"
asgiref = "                      3.8.1"
astroid = "                      3.2.2"
bcrypt = "                       4.1.3"
billiard = "                     4.2.0"
black = "                        24.4.2"
celery = "                       5.4.0"
certifi = "                      2024.7.4"
cffi = "                         1.16.0"
charset-normalizer = "           3.3.2"
click = "                        8.1.7"
click-didyoumean = "             0.3.1"
click-plugins = "                1.1.1"
click-repl = "                   0.3.0"
colorama = "                     0.4.6"
cryptography = "                 42.0.8"
dill = "                         0.3.8"
dj-database-url = "              2.2.0"
django = "                       5.0.6"
django-cors-headers = "          4.3.1"
django-environ = "               0.11.2"
django-extensions = "            3.2.3"
djangorestframework = "          3.15.1"
djangorestframework-api-key = "  3.0.0"
djangorestframework-simplejwt = "5.3.1"
drf-yasg = "                     1.21.7"
gunicorn = "                     22.0.0"
idna = "                         3.7"
inflection = "                   0.5.1"
isort = "                        5.13.2"
kombu = "                        5.3.7"
mccabe = "                       0.7.0"
mypy-extensions = "              1.0.0"
packaging = "                    24.0"
paramiko = "                     3.4.0"
pathspec = "                     0.12.1"
pillow = "                       10.3.0"
pip = "                          24.0"
platformdirs = "                 4.2.2"
prompt-toolkit = "               3.0.47"
psycopg2 = "                     2.9.9"
pycparser = "                    2.22"
pyjwt = "                        2.8.0"
pylint = "                       3.2.2"
pynacl = "                       1.5.0"
python-dateutil = "              2.9.0.post0"
python-decouple = "              3.8"
python-dotenv = "                1.0.1"
pytz = "                         2024.1"
pywin32 = "                      306"
pyyaml = "                       6.0.1"
redis = "                        5.0.7"
requests = "                     2.32.3"
setuptools = "                   70.0.0"
six = "                          1.16.0"
sqlparse = "                     0.5.0"
sshtunnel = "                    0.4.0"
tomlkit = "                      0.12.5"
typing-extensions = "            4.12.2"
tzdata = "                       2024.1"
uritemplate = "                  4.1.1"
urllib3 = "                      2.2.2"
vine = "                         5.1.0"
wcwidth = "                      0.2.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
